+++ src/main/java/models/TimeHandler.java
@@ -27,4 +27,27 @@ public class TimeHandler {
 		return finishTime.get(i);
 	}
 
+	public String print() {
+		StringBuilder sb = new StringBuilder();
+		sb.append("Start Time; Finish Time\n");
+		if (startTime.size() == 0) {
+			if (finishTime.size() == 0) {
+				return "Empty time list.";
+			}
+			for (String s : finishTime) {
+				sb.append("--.--.--; " + s + "\n");
+			}
+			return sb.toString();
+		} else if (finishTime.size() == 0) {
+			for (String s : startTime) {
+				sb.append(s + "; --.--.--\n");
+			}
+			return sb.toString();
+		} else {
+			for (int i = 0; i < startTime.size(); i++) {
+				sb.append(startTime.get(i) + "; " + finishTime.get(i) + "\n");
+			}
+			return sb.toString();
+		}
+	}
 }
