+++ src/test/java/models/PrintTest.java
@@ -8,17 +8,18 @@ import org.junit.Test;
 
 public class PrintTest {
 
-	TimeHandlerDummy th;
+	TimeHandler th;
+
 	@Before
 	public void setUp() {
-		th = new TimeHandlerDummy();
+		th = new TimeHandler();
 	}
 
 	@Test
 	public void testAddThenPrint() {
 		th.addStart("12.12.12");
 		th.addFinish("13.13.13");
-		assertEquals(th.print(), "12.12.12; 13.13.13");
+		assertEquals(th.print(), "Start Time; Finish Time\n12.12.12; 13.13.13\n");
 	}
 
 	@Test
@@ -30,12 +31,12 @@ public class PrintTest {
 	@Test
 	public void testPrintMissingFinishTime() {
 		th.addStart("12.12.12");
-		assertEquals(th.print(), "12.12.12; --.--.--");
+		assertEquals(th.print(), "Start Time; Finish Time\n12.12.12; --.--.--\n");
 	}
 
 	public void testPrintMissingStartTime() {
 		th.addFinish("12.12.12");
-		assertEquals(th.print(), "--.--.--; 12.12.12");
+		assertEquals(th.print(), "Start Time; Finish Time\n--.--.--; 12.12.12\n");
 	}
 
 	@Test
@@ -47,6 +48,6 @@ public class PrintTest {
 		th.addFinish("13.13.13");
 		th.addFinish("13.13.14");
 		assertEquals(th.print(),
-				"12.12.12; 13.13.12\n12.12.13; 13.13.13\n12.12.14; 13.13.14");
+				"Start Time; Finish Time\n12.12.12; 13.13.12\n12.12.13; 13.13.13\n12.12.14; 13.13.14\n");
 	}
 }
